refactor: share phone verification logic
