+++ src/shared/utils/format-remaining-time.ts
