+++ src/shared/utils/normalize-phone-number.ts
